Update MSR prototype, PMM counters & bitmap ops
Change KrWriteModelSpecificRegister to accept a QWORD (header and implementation) for correct 64-bit MSR writes. In KrKernelStart, add Virtmemmgmt include and a linker-symbol comment. In Physmemmgmt: remove unused includes, increment/decrement kernel PMM counters when pages are acquired/relinquished/reserved, fix a for-loop initialization, and replace byte-by-byte bulk bitmap writes with staged 8/4/2/1-byte writes for better performance.

+++ Kernel/Include/CPU/MSR.h
@@ -6,6 +6,6 @@
 // `RDMSR`
 QWORD KrReadModelSpecificRegister(DWORD idRegister);
 // `WRMSR`
-void  KrWriteModelSpecificRegister(DWORD idRegister, DWORD data);
+void  KrWriteModelSpecificRegister(DWORD idRegister, QWORD data);
 
 #endif // !YCH_KERNEL_CPU_MSR_H
